添加拆解方法
