+++ TextTool/TextTool.csproj
@@ -59,6 +59,7 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TextHandler.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="System" />
